+++ changelog.md
@@ -1,4 +1,11 @@
 
+## Version 3.20 (2014-10-18)
+
+1. **Added Tags Export Preferences**
+
+    Added Tags Export application preferences, allowing the user to specify tags to select for export, and/or tags to suppress.
+
+
 ## Version 3.10 (2013-04-23)
 
 1. **Released as Open Source Software**